# Use k-means to construct hyperplanes