--- a/.travis.yml
+++ b/.travis.yml
@@ -12,5 +12,5 @@ install:
   - pip install tox
 
 script:
-  - tox
+  - TRAVIS=true tox