--- a/src/annoylib.h
+++ b/src/annoylib.h
@@ -88,6 +88,42 @@ inline void normalize(T* v, int f) {
     v[z] /= norm;
 }
 
+template<typename T, typename Random, typename Distance, typename Node>
+inline void two_means(const vector<Node*>& nodes, int f, Random& random, bool cosine, T* iv, T* jv) {
+  /*
+    This algorithm is a huge heuristic. Empirically it works really well, but I
+    can't motivate it well. The basic idea is to keep two centroids and assign
+    points to either one of them. We weight each centroid by the number of points
+    assigned to it, so to balance it. 
+  */
+  static int iteration_steps = 200;
+  size_t count = nodes.size();
+
+  size_t i = random.index(count);
+  size_t j = random.index(count-1);
+  j += (j >= i); // ensure that i != j
+  std::copy(&nodes[i]->v[0], &nodes[i]->v[f], &iv[0]);
+  std::copy(&nodes[j]->v[0], &nodes[j]->v[f], &jv[0]);
+  if (cosine) { normalize(&iv[0], f); normalize(&jv[0], f); }
+
+  int ic = 1, jc = 1;
+  for (int l = 0; l < iteration_steps; l++) {
+    size_t k = random.index(count);
+    T di = ic * Distance::distance(&iv[0], nodes[k]->v, f),
+      dj = jc * Distance::distance(&jv[0], nodes[k]->v, f);
+    T norm = cosine ? get_norm(nodes[k]->v, f) : 1.0;
+    if (di < dj) {
+      for (int z = 0; z < f; z++)
+	iv[z] = (iv[z] * ic + nodes[k]->v[z] / norm) / (ic + 1);
+      ic++;
+    } else if (dj < di) {
+      for (int z = 0; z < f; z++)
+	jv[z] = (jv[z] * jc + nodes[k]->v[z] / norm) / (jc + 1);
+      jc++;
+    }
+  }
+}
+
 
 struct Angular {
   template<typename S, typename T>
@@ -142,19 +178,10 @@ struct Angular {
   }
   template<typename S, typename T, typename Random>
   static inline void create_split(const vector<Node<S, T>*>& nodes, int f, Random& random, Node<S, T>* n) {
-    // Sample two random points from the set of nodes
-    // Calculate the hyperplane equidistant from them
-    size_t count = nodes.size();
-    size_t i = random.index(count);
-    size_t j = random.index(count-1);
-    j += (j >= i); // ensure that i != j
-    T* iv = nodes[i]->v;
-    T* jv = nodes[j]->v;
-    T i_norm = get_norm(iv, f);
-    T j_norm = get_norm(jv, f);
+    static vector<T> best_iv(f, 0), best_jv(f, 0);
+    two_means<T, Random, Angular, Node<S, T> >(nodes, f, random, true, &best_iv[0], &best_jv[0]);
     for (int z = 0; z < f; z++)
-      n->v[z] = iv[z] / i_norm - jv[z] / j_norm;
-    normalize(n->v, f);
+      n->v[z] = best_iv[z] - best_jv[z];
   }
   template<typename T>
   static inline T normalized_distance(T distance) {
@@ -196,20 +223,16 @@ struct Euclidean {
       return random.flip();
   }
   template<typename S, typename T, typename Random>
-  static inline void create_split(const vector<Node<S, T>*>& nodes, int f, Random& random, Node<S, T>* n) {
-    // Same as Angular version, but no normalization and has to compute the offset too
-    size_t count = nodes.size();
-    size_t i = random.index(count);
-    size_t j = random.index(count-1);
-    j += (j >= i); // ensure that i != j
-    T* iv = nodes[i]->v;
-    T* jv = nodes[j]->v;
+    static inline void create_split(const vector<Node<S, T>*>& nodes, int f, Random& random, Node<S, T>* n) {
+    static vector<T> best_iv(f, 0), best_jv(f, 0);
+    two_means<T, Random, Euclidean, Node<S, T> >(nodes, f, random, false, &best_iv[0], &best_jv[0]);
+
     for (int z = 0; z < f; z++)
-      n->v[z] = iv[z] - jv[z];
+      n->v[z] = best_iv[z] - best_jv[z];
     normalize(n->v, f);
     n->a = 0.0;
     for (int z = 0; z < f; z++)
-      n->a += -n->v[z] * (iv[z] + jv[z]) / 2;
+      n->a += -n->v[z] * (best_iv[z] + best_jv[z]) / 2;
   }
   template<typename T>
   static inline T normalized_distance(T distance) {
@@ -448,47 +471,36 @@ protected:
     Node* m = (Node*)malloc(_s); // TODO: avoid
 
     vector<S> children_indices[2];
-    for (int attempt = 0; attempt < 20; attempt ++) {
-      /*
-       * Create a random hyperplane.
-       * If all points end up on the same time, we try again.
-       * We could in principle *construct* a plane so that we split
-       * all items evenly, but I think that could violate the guarantees
-       * given by just picking a hyperplane at random
-       */
-      vector<Node*> children;
 
-      for (size_t i = 0; i < indices.size(); i++) {
-        // TODO: this loop isn't needed for the angular distance, because
-        // we can just split by a random vector and it's fine. For Euclidean
-        // distance we need it to calculate the offset
-        S j = indices[i];
-        Node* n = _get(j);
-        if (n)
-          children.push_back(n);
-      }
+    // Create a random hyperplane.
+    vector<Node*> children;
 
-      D::create_split(children, _f, _random, m);
+    for (size_t i = 0; i < indices.size(); i++) {
+      // TODO: this loop isn't needed for the angular distance, because
+      // we can just split by a random vector and it's fine. For Euclidean
+      // distance we need it to calculate the offset
+      S j = indices[i];
+      Node* n = _get(j);
+      if (n)
+        children.push_back(n);
+    }
 
-      children_indices[0].clear();
-      children_indices[1].clear();
+    D::create_split(children, _f, _random, m);
 
-      for (size_t i = 0; i < indices.size(); i++) {
-        S j = indices[i];
-        Node* n = _get(j);
-        if (n) {
-          bool side = D::side(m, n->v, _f, _random);
-          children_indices[side].push_back(j);
-        }
-      }
+    children_indices[0].clear();
+    children_indices[1].clear();
 
-      if (children_indices[0].size() > 0 && children_indices[1].size() > 0) {
-        break;
+    for (size_t i = 0; i < indices.size(); i++) {
+      S j = indices[i];
+      Node* n = _get(j);
+      if (n) {
+        bool side = D::side(m, n->v, _f, _random);
+        children_indices[side].push_back(j);
       }
     }
 
+    // If we didn't find a hyperplane, just randomize sides as a last option
     while (children_indices[0].size() == 0 || children_indices[1].size() == 0) {
-      // If we didn't find a hyperplane, just randomize sides as a last option
       if (_verbose && indices.size() > 100000)
         showUpdate("Failed splitting %lu items\n", indices.size());
 
@@ -568,7 +580,7 @@ protected:
     std::partial_sort(&nns_dist[0], &nns_dist[p], &nns_dist[m]);
     for (size_t i = 0; i < p; i++) {
       if (distances)
-	distances->push_back(D::normalized_distance(nns_dist[i].first));
+        distances->push_back(D::normalized_distance(nns_dist[i].first));
       result->push_back(nns_dist[i].second);
     }
   }