--- a/tox.ini
+++ b/tox.ini
@@ -2,6 +2,8 @@
 envlist=py{26,27,33,34}, go
 
 [testenv]
+setenv =
+  TRAVIS = {env:TRAVIS:}
 commands =
   pip install numpy
   pip install .