This seems like a pretty huge improvement according to some benchmarks I ran.

```
Euclidean, f=100, GloVe: 18.211s -> 16.573s, 89.10% -> 95.98%
Angular,   f=100, GloVe: 38.947s -> 32.799s, 89.07% -> 96.74%
```

Only drawback is tree building is marginally slower