@a1k0n 2015-10-12
Oh, there are mixed tabs and spaces, which explains it. Heh.

Anyways, that is awesome. And sort of obvious in retrospect.