@erikbern 2015-10-12
i have a couple of smaller commits coming up, so don't merge yet.

yeah this is quite obvious and i noticed last time i tried it i had a couple of bugs in my code. doing the math i think this is roughly a 2x speedup (i.e. hopefully now it should beat all other ANN libraries)