@a1k0n 2015-10-12
I dunno, the small world graph type algos have pretty convincing results...