@erikbern 2015-10-12
SW-graph (nmslib) and kgraph are less than 2x faster than Annoy so I hope it's doable :)