@a1k0n 2015-10-12
You still aren't updating best_d_sum on line 134 (apparently github mobile doesn't let me comment on lines directly)