@erikbern 2015-10-13
Some benchmarks below.

The problem is I'm getting segfaults when trying to use 400 trees. Fixing this would give even better results since now I'm maxing out at 200 trees. The k-means algorithm uses a bunch more dynamic memory allocation so I will try to rewrite the tree building algorithm to use as little as possible.

Tree building is also 10x slower which is a bit unfortunate but Annoy still is much faster than other algos

![image](https://cloud.githubusercontent.com/assets/1027979/10453913/600165c6-717e-11e5-8a95-a4e0b49ebe76.png)