@erikbern 2015-10-13
Turns out since I'm subsampling it's better to do it using SGD rather than in batches. Also seems like the number of attempts doesn't matter at that point – better to increase the number of samples. So code is much simpler now.