@a1k0n 2015-10-13
Nice, that makes a lot of sense. Looks good.