@a1k0n 2015-11-17
@erikbern Erik, are you going to merge this? It seems like a good idea and it seems to work better..?