@erikbern 2015-11-17
i'm planning to merge this but i want to spend a bit more time first – the reason is that I don't fully understand why it works so well, so I'm a bit concerned it will introduce regressions for certain cases.