@erikbern 2016-01-31
rebased this, will probably merge this after a bit of cleanup and a few sanity tests to make sure rebasing didn't mess up anything